backend/src/initialize_db.py
from db import Base, engine
import dotenv
dotenv.load_dotenv()

def init_db():
    print("Creating database tables...")
    Base.metadata.create_all(bind=engine)
    print("Database tables created successfully.")

if __name__ == "__main__":
    init_db()
