backend/requirements.txt
aiohttp==3.11.7
cryptography==43.0.3
fastapi==0.115.5
passlib==1.7.4
pydantic==2.10.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
uvicorn==0.32.1
